requirements.txt
PyYAML==6.0
psutil==5.8.0